+++ src/main/java/uk/gov/hmcts/cmc/claimstore/ClaimStoreApplication.java
@@ -5,13 +5,13 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
 import org.springframework.cloud.netflix.feign.EnableFeignClients;
 
-@SpringBootApplication(exclude = DataSourceAutoConfiguration.class)
+@SpringBootApplication(exclude = DataSourceAutoConfiguration.class, scanBasePackages = "uk.gov.hmcts")
 @SuppressWarnings({"HideUtilityClassConstructor", "squid:S1118"}) // Spring needs a constructor, its not a utility class
 @EnableFeignClients(basePackages =
     {"uk.gov.hmcts.cmc.claimstore",
         "uk.gov.hmcts.document",
         "uk.gov.hmcts.reform.authorisation",
-        "uk.gov.hmcts.cmc.ccd"
+        "uk.gov.hmcts.reform.ccd"
     })
 public class ClaimStoreApplication {
