+++ src/main/java/uk/gov/hmcts/cmc/claimstore/services/ccd/CoreCaseDataService.java
@@ -3,11 +3,11 @@ package uk.gov.hmcts.cmc.claimstore.services.ccd;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.stereotype.Service;
-import uk.gov.hmcts.cmc.ccd.client.model.CaseDetails;
-import uk.gov.hmcts.cmc.ccd.client.model.EventRequestData;
 import uk.gov.hmcts.cmc.ccd.domain.CCDCase;
 import uk.gov.hmcts.cmc.ccd.mapper.CaseMapper;
 import uk.gov.hmcts.cmc.domain.models.Claim;
+import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
+import uk.gov.hmcts.reform.ccd.client.model.EventRequestData;
 
 @Service
 @ConditionalOnProperty(prefix = "feature_toggles", name = "core_case_data", havingValue = "true")
@@ -21,7 +21,10 @@ public class CoreCaseDataService {
     private final CaseMapper caseMapper;
 
     @Autowired
-    public CoreCaseDataService(final SaveCoreCaseDataService saveCoreCaseDataService, final CaseMapper caseMapper) {
+    public CoreCaseDataService(
+        final SaveCoreCaseDataService saveCoreCaseDataService,
+        final CaseMapper caseMapper
+    ) {
         this.saveCoreCaseDataService = saveCoreCaseDataService;
         this.caseMapper = caseMapper;
     }
