+++ src/main/resources/application.yml
@@ -87,7 +87,7 @@ staff-notifications:
 feature_toggles:
   watermark_pdf: false
   document_management: false
-  core_case_data: false
+  core_case_data: true
 
 feign:
   httpclient:
